refactor : methods, errors, responses, etc.

+++ src/main/java/ar/edu/unq/desapp/grupoG/backendapicryptoexchange/API/contracts/TransactionIntention/CreateTransactionIntentionResponse.java
@@ -8,7 +8,7 @@ import lombok.Data;
 import java.time.LocalDateTime;
 
 public record CreateTransactionIntentionResponse (
-    Integer transaction_intention_id,
+    Long transaction_intention_id,
     String operation_type,
     String crypto_symbol,
     Double crypto_price,
